Adds LDF saving feature (#90)
LDF object can be saved using a template
Lark parser version bumped to first major version
Transformers have been moved into their own file
Fixes some Node attributes defaulting to None
Adds Jinja2 as a requirement
Signal encoders are now linked to the signals
LDF has getters for signal encoders
Update readme with the new feature
Bumps version to 0.14.0
Fixed missing node definition in a test LDF file
Adds flake8 problem matcher to build workflow
