+++ .github/workflows/build.yml
@@ -64,6 +64,9 @@ jobs:
       - name: Checkout
         uses: actions/checkout@v2
 
+      - name: Setup flake8 annotations
+        uses: rbialon/flake8-annotations@v1
+
       - name: Setup
         run: |
           python -m pip install --upgrade pip
