+++ setup.cfg
@@ -1,2 +1,2 @@
 [metadata]
-version = 0.13.1
+version = 0.14.0
