+++ tests/ldf/no_signal_subscribers.ldf
@@ -27,6 +27,11 @@ Frames {
 
 Node_attributes {
     // Empty
+    slave1 {
+        LIN_protocol = 2.0;
+        configured_NAD = 0x01;
+        product_id = 0x4A4F, 0x4841;
+    }
 }
 
 Schedule_tables {
