+++ tests/test_parser.py
@@ -159,10 +159,13 @@ def test_load_valid_lin_encoders():
     assert len(converter._converters) == 1
     assert isinstance(converter._converters[0], BCDValue)
 
-    converter = ldf.converters['ascii_signal']
+    converter = ldf.get_signal_encoding_type('AsciiEncoding')
     assert converter.name == 'AsciiEncoding'
     assert len(converter._converters) == 1
     assert isinstance(converter._converters[0], ASCIIValue)
+
+    with pytest.raises(LookupError):
+        ldf.get_signal_encoding_type('abc123')
 
 @pytest.mark.unit
 def test_load_valid_diagnostics():
